8.77

+++ app/src/main/java/ru/kavunov/mtsproject/MovieClickListener.kt
@@ -3,7 +3,7 @@ package ru.kavunov.mtsproject
 import android.content.Context
 
 interface MovieClickListener {
-    fun clickDetail(position: Int)
+    fun clickDetail(position: Long)
 
 
 }
